+++ core/src/main/java/com/shatteredpixel/yasd/items/wands/WandOfFrost.java
@@ -45,7 +45,7 @@ public class WandOfFrost extends DamageWand {
 		image = ItemSpriteSheet.WAND_FROST;
 	}
 
-	public int min(int lvl){
+	public float min(float lvl){
 		return 2+lvl;
 	}
